+++ components/layout/layout.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import Head from "next/head";
 import { Header } from "./header";
+import { FooterNav } from "./footer-nav";
 import { Blocks } from "../../components/blocks-renderer";
 import { styles } from "./styles"
 import { googleFontsLink } from "./google-fonts"
@@ -71,6 +72,9 @@ export const Layout = ({ rawData, children }) => {
         {children}
       </main>
       <div id="footer">
+        { global.footerNav === "true" && (
+          <FooterNav blocks={page?.blocks} globalData={global} />
+        )}
         <Blocks blocks={global.blocks} />
       </div>
     </>