+++ tina/config.tsx
@@ -6,7 +6,7 @@ import { backgroundSchema } from "../schema/background"
 // Blocks
 import { featureBlockSchema } from "../schema/blocks/feature";
 import { cardsBlockSchema } from "../schema/blocks/cards";
-import { accordianBlockSchema } from "../schema/blocks/accordian";
+import { accordionBlockSchema } from "../schema/blocks/accordion";
 import { embedBlockSchema } from "../schema/blocks/embed";
 import { tailwindFeatureBlockSchema } from "../schema/blocks/tailwind-feature"
 import { tailwindCardsBlockSchema } from "../schema/blocks/tailwind-cards"
@@ -79,7 +79,7 @@ const config = defineStaticConfig({
             templates: [
               featureBlockSchema,
               cardsBlockSchema,
-              accordianBlockSchema,
+              accordionBlockSchema,
               embedBlockSchema,
               tailwindFeatureBlockSchema,
               tailwindCardsBlockSchema,